make a clear division between the the service layer and the model and view layers
because that is more readable, maintainable and testable
