+++ src/service/index.ts
@@ -1,9 +1,11 @@
+import * as auth from "./auth";
 import * as search from "./search";
 import * as user from "./user";
 
 export const service = {
   search,
   user,
+  auth,
 };
 
 export function mockService(record: Record<string, any> = service) {
